Add (Partial )Day #4 work

+++ Squeaky Clicker/main.py
@@ -41,21 +41,21 @@ item_font = pygame.font.SysFont('Comic Sans MS', 30)
 floaty_point_text = pygame.font.SysFont('Comic Sans MS', 25, bold=True)
 menu_header = points_font
 additional_number = random.randint(1, 10)
-#Main Game LooLoop
+#Main Game Loop
 while running:
     dog_cost = dog_level+1
     dog_cost *= 35
     dog_cost **= 1.5
     dog_cost += additional_number
     dog_cost = int(dog_cost)
-    squeaker_cost = squeaker_level*2
+    squeaker_cost = squeaker_level*20
     squeaker_cost **= 1.5
     squeaker_cost += additional_number
     squeaker_cost = int(squeaker_cost)
     screen.fill((255, 255, 255))
     for event in pygame.event.get():
         if event.type == pygame.QUIT:
-            if os.path.exists("./saves/save.json"):
+            if os.path.exists("./saves"):
                 with open("./saves/save.json", "w") as f:
                     data = {
                         "squeaky": squeaker_level,
@@ -71,19 +71,24 @@ while running:
             running = False
         if event.type == pygame.KEYDOWN:
             if event.key == pygame.K_SPACE:
-                points_tobeadded = 1*squeaker_level
+                points_tobeadded = squeaker_level
+                golden_squeak_chance = random.randint(1, 472)
+                if golden_squeak_chance == 360:
+                    points_tobeadded *= 100
                 points += points_tobeadded
                 squeak = pygame.mixer.Sound('./assets/squeakers/squeak.wav')
                 squeak.play()
                 floaty_point_text_text = floaty_point_text.render(f"+{points_tobeadded}", True, (0, 0, 0))
-                screen.blit(floaty_point_text_text, (400, 100))
+                floaty_text_x = 0.1*points
+                screen.blit(floaty_point_text_text, (400+floaty_text_x, 100))
             if event.key == pygame.K_0:
                 #Yes i know this is a bad name ill change it later.
                 tmpvar = squeaker_level *20
                 points_toberemoved = tmpvar**1.5
+                print(f"Points: {points}, Points to be removed: {points_toberemoved}")
                 if int(points) >= points_toberemoved:
                     print(f"Points: {points}, Points to be removed: {points_toberemoved}")
-                    squeaker_level += 1
+                    squeaker_level *= 1.2
                     print(points_toberemoved)
                     points -= points_toberemoved+additional_number
                     additional_number = random.randint(1, 10)
@@ -130,14 +135,20 @@ while running:
             pygame.display.flip()
             logo_shown = True
     if not menu_shown:
+        has_reset = False
         squeaked = False
+        reset = ""
         while not squeaked:
+            screen.fill((0, 0, 0))
             header = menu_header.render(f"Squeaky Clicker", True, (255, 255, 255))
             screen.blit(header, (500, 100))
             playtext = menu_header.render(f"Squeak to begin...", True, (255, 255, 255))
             reset_text = menu_header.render(f"Press down on 1 to reset save data", True, (255, 255, 255))
             screen.blit(playtext, (500, 200))
             screen.blit(reset_text, (500, 300))
+            if has_reset:
+                reset_result = menu_header.render(reset, True, (255, 255, 255))
+                screen.blit(reset_result, (500, 700))
             pygame.display.flip()
             for event in pygame.event.get():
                 for event in pygame.event.get():
@@ -147,13 +158,40 @@ while running:
                     if event.key == pygame.K_SPACE:
                         squeaked = True
                         menu_shown = True
+                        squeak = pygame.mixer.Sound("./assets/squeakers/squeak.wav")
+                        squeak.play()
+                        for i in range(255):
+                            screen.fill((i, i, i))
+                            pygame.display.flip()
+                            time.sleep(0.007)
                     if event.key == pygame.K_1:
-                        os.remove("./saves/save.json")
-                        squeaked = True
-                        menu_shown = True
+                        while not has_reset:
+                            screen.fill((0, 0, 0))
+                            reset_confirm = menu_header.render("Are you sure you want to reset your save data?", True, (255, 255, 255))
+                            reset_confirm_2 = menu_header.render("Squeak to proceed.", True, (255, 255, 255))
+                            reset_escape = menu_header.render("Press down on 1 to cancel.", True, (255, 255, 255))
+                            screen.blit(reset_confirm, (500, 200))
+                            screen.blit(reset_confirm_2, (500, 300))
+                            screen.blit(reset_escape, (500, 400))
+                            pygame.display.flip()
+                            for event in pygame.event.get():
+                                if event.type == pygame.QUIT:
+                                    running = False
+                                    has_reset = False
+                                if event.type == pygame.KEYDOWN:
+                                    if event.key == pygame.K_SPACE:
+                                        try:
+                                            os.remove("./saves/save.json")
+                                        except:
+                                            pass
+                                        has_reset = True
+                                        reset = "Reset Successful!"
+                                    if event.key == pygame.K_1:
+                                        has_reset = True
+                                        reset = ""
     x = random.randint(1, 2)
     if x == 1:
-        points += dog_level*0.01
+        points += squeaker_level*dog_level*0.01
     else:
         pass
     points_surface = points_font.render(f'Squeaks: {int(points)}', True, (0, 0, 0))
